+++ packages/picasso-lab/src/NoteSubtitle/NoteSubtitle.tsx
@@ -10,7 +10,7 @@ export const NoteSubtitle = forwardRef<HTMLDivElement, Props>(
   function NoteSubtitle({ children, ...rest }, ref) {
     return (
       <Container ref={ref} bottom='small' {...rest}>
-        <Typography color='dark-grey' size='small'>
+        <Typography color='dark-grey' size='xsmall'>
           {children}
         </Typography>
       </Container>