+++ packages/picasso/src/PageTopBarMenu/PageTopBarMenu.tsx
@@ -35,7 +35,7 @@ export const PageTopBarMenu = forwardRef<HTMLDivElement, Props>(
 
     const metaContent =
       typeof meta === 'string' ? (
-        <Typography className={classes.truncateText} invert size='small'>
+        <Typography className={classes.truncateText} invert size='xsmall'>
           {meta}
         </Typography>
       ) : (