+++ packages/picasso/src/Radio/story/CustomLabel.example.tsx
@@ -33,12 +33,12 @@ const Benefits = () => {
     <Container>
       <List>
         <li>
-          <Typography size='small'>
+          <Typography size='xsmall'>
             Similar benefits to Toptal Payments
           </Typography>
         </li>
         <li>
-          <Typography size='small'>
+          <Typography size='xsmall'>
             Instantly transfer funds to a prepaid card{' '}
           </Typography>
         </li>
@@ -71,13 +71,7 @@ const PayoneerPicker = () => {
 
 const Example = () => (
   <div>
-    <Container
-      bordered
-      rounded
-      padded='medium'
-      variant='blue'
-      style={{ maxWidth: 400 }}
-    >
+    <Container rounded padded='medium' variant='blue' style={{ maxWidth: 400 }}>
       <PayoneerPicker />
       <Benefits />
     </Container>