+++ packages/picasso/src/TableCell/TableCell.tsx
@@ -1,12 +1,19 @@
 import React, { forwardRef, HTMLAttributes, useContext } from 'react'
 import { makeStyles, Theme } from '@material-ui/core/styles'
 import MUITableCell from '@material-ui/core/TableCell'
-import { BaseProps, TextLabelProps, useTitleCase } from '@toptal/picasso-shared'
+import {
+  BaseProps,
+  ColorType,
+  SizeType,
+  TextLabelProps,
+  useTitleCase
+} from '@toptal/picasso-shared'
 import cx from 'classnames'
 
 import toTitleCase from '../utils/to-title-case'
 import styles from './styles'
 import { TableContext, TableSectionContext, TableSection } from '../Table'
+import Typography from '../Typography'
 
 type AlignType = 'inherit' | 'left' | 'center' | 'right' | 'justify'
 
@@ -24,6 +31,36 @@ export interface Props
 
 const useStyles = makeStyles<Theme>(styles, { name: 'PicassoTableCell' })
 
+type TypographySettings = {
+  color: ColorType
+  size: SizeType<'xsmall' | 'small'>
+  weight: 'regular' | 'semibold' | 'inherit'
+}
+const getTypographySettings = (
+  tableSection: TableSection | undefined
+): TypographySettings => {
+  switch (tableSection) {
+    case TableSection.HEAD:
+      return {
+        color: 'dark-grey',
+        size: 'xsmall',
+        weight: 'semibold'
+      }
+    case TableSection.FOOTER:
+      return {
+        color: 'black',
+        size: 'small',
+        weight: 'semibold'
+      }
+    default:
+      return {
+        color: 'dark-grey',
+        size: 'small',
+        weight: 'inherit'
+      }
+  }
+}
+
 export const TableCell = forwardRef<HTMLTableCellElement, Props>(
   function TableCell(props, ref) {
     const {
@@ -39,15 +76,12 @@ export const TableCell = forwardRef<HTMLTableCellElement, Props>(
 
     const {
       compact: compactClass,
-      footer: footerClass,
-      header: headerClass,
       narrow: narrowClass,
       ...muiClasses
     } = useStyles()
     const { spacing } = useContext(TableContext)
     const tableSection = useContext(TableSectionContext)
     const isHead = tableSection === TableSection.HEAD
-    const isFooter = tableSection === TableSection.FOOTER
     const titleCase = useTitleCase(propsTitleCase)
 
     const renderChildren = () =>
@@ -61,15 +95,15 @@ export const TableCell = forwardRef<HTMLTableCellElement, Props>(
         classes={muiClasses}
         className={cx(className, {
           [compactClass]: spacing === 'compact',
-          [narrowClass]: spacing === 'narrow',
-          [footerClass]: isFooter,
-          [headerClass]: isHead
+          [narrowClass]: spacing === 'narrow'
         })}
         style={style}
         colSpan={colSpan}
         rowSpan={rowSpan}
       >
-        {renderChildren()}
+        <Typography as='div' {...getTypographySettings(tableSection)}>
+          {renderChildren()}
+        </Typography>
       </MUITableCell>
     )
   }