+++ packages/picasso/src/TableSectionHead/styles.ts
@@ -1,10 +1,11 @@
 import { createStyles, Theme } from '@material-ui/core/styles'
 import { rem } from '@toptal/picasso-shared'
 
-export default ({ palette, sizes, typography }: Theme) =>
+export default ({ palette, sizes }: Theme) =>
   createStyles({
     iconWrapper: {
       marginRight: '0.75rem',
+      marginTop: rem('-1px'),
       display: 'inline-block',
       verticalAlign: 'middle'
     },
@@ -12,10 +13,5 @@ export default ({ palette, sizes, typography }: Theme) =>
       minHeight: '2.5rem',
       borderBottom: `${sizes.borderWidth} solid ${palette.grey.lighter2}`,
       borderTop: `${sizes.borderWidth} solid ${palette.grey.lighter2}`
-    },
-    sectionHeaderCell: {
-      fontSize: rem('12px'),
-      fontWeight: typography.fontWeights.semibold,
-      color: palette.text.primary
     }
   })